image_similarity_EfficientNet/requirements.txt
aniso8601==9.0.1
attrs==21.4.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cycler==0.11.0
dataclasses==0.8
flasgger==0.9.5
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-RESTful==0.3.9
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==2.0.2
numpy==1.19.5
opencv-python==4.5.5.62
Pillow==8.4.0
pycparser==2.21
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
torch==1.10.2
torchaudio==0.10.2
torchvision==0.11.3
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
